task-manager-api/app/models.py
from tortoise import fields
from tortoise.models import Model


class Task(Model):
    task_id = fields.CharField(pk=True, max_length=36)
    user_id = fields.CharField(max_length=36)
    title = fields.CharField(max_length=255)
    description = fields.TextField()
    completed = fields.BooleanField(default=False)
