task-manager-api/requirements.txt
fastapi
uvicorn
boto3
tortoise-orm
pydantic
